Rename each types from type
