+++ src/characters/fighter/createFighter.test.ts
@@ -1,10 +1,11 @@
+import { mandatoryCharacterData } from "../character/types";
 import { createFighter } from "./createFighter";
-import { DexterityRange, Fighter } from "./type";
+import { DexterityRange, Fighter } from "./types";
 
 describe("Given the function createFighter", () => {
   describe("When it receives the name Jaime, the surname Lannister and a sword as a weapon", () => {
     test("Then it should return a fighter named Jaime Lannister that uses a sword as a weapon", () => {
-      const mandatoryCharacterData = {
+      const mandatoryCharacterData: mandatoryCharacterData = {
         name: "Jaime",
         surname: "Lannister",
         age: 36,
