+++ src/characters/fighter/types.ts
@@ -1,4 +1,4 @@
-import { Character } from "../../types";
+import { Character } from "../character/types";
 
 export type DexterityRange = 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10;
