+++ src/characters/king/createKing.test.ts
@@ -1,11 +1,11 @@
-import { mandatoryData } from "../../types";
+import { mandatoryCharacterData } from "../character/types";
 import { createKing } from "./createKing";
-import { King } from "./type";
+import { King } from "./types";
 
 describe("Given the function createKing", () => {
   describe("When it receives the name Joffrey, the surname Baratheon and the age of 16", () => {
     test("Then it should return a king named Joffrey Baratheon who has the age of 16", () => {
-      const mandatoryCharacterData: mandatoryData = {
+      const mandatoryCharacterData: mandatoryCharacterData = {
         name: "Joffrey",
         surname: "Baratheon",
         age: 16,
