+++ src/characters/squire/createSquire.test.ts
@@ -1,11 +1,15 @@
-import { Fighter } from "../fighter/type";
+import { mandatoryCharacterData } from "../character/types";
+import { Fighter } from "../fighter/types";
 import { createSquire } from "./createSquire";
-import { FidelityRange } from "./type";
+import { FidelityRange } from "./types";
 
 describe("Given the function createSquire", () => {
   describe("When it receives the name Bronn and the serving person's name Jaime", () => {
     test("Then it should return a squire named Bronn who serves someone named Jaime", () => {
-      const mandatoryCharacterData = { name: "Bronn", age: 25 };
+      const mandatoryCharacterData: mandatoryCharacterData = {
+        name: "Bronn",
+        age: 25,
+      };
       const expectedSquireName = "Bronn";
       const expectedServedCharacterName = "Jaime";
       const fidelity: FidelityRange = 7;
