+++ src/characters/squire/types.ts
@@ -1,5 +1,5 @@
-import { Character } from "../../types";
-import { Fighter } from "../fighter/type";
+import { Character } from "../character/types";
+import { Fighter } from "../fighter/types";
 
 export type FidelityRange = 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10;
