--- src/types.ts
@@ -1,11 +1,0 @@
-export interface mandatoryData {
-  name: string;
-  surname?: string;
-  age: number;
-}
-
-export interface Character extends mandatoryData {
-  isAlive: boolean;
-  die: () => void;
-  talk: () => string;
-}
